+++ test/Projects/index.test.js
@@ -1,5 +1,10 @@
 import Projects from "../../src/Projects";
-import { testGetter, testPost, testUpdate } from "../utils/http";
+import {
+  testGetter,
+  testParamsGetter,
+  testPost,
+  testUpdate,
+} from "../utils/http";
 
 describe("Projects#all", () => {
   testGetter(
@@ -57,6 +62,41 @@ describe("Projects#update", () => {
       return new Projects(client).update("project-id", {
         name: "new-name",
         default: true,
+      });
+    },
+  );
+});
+
+describe("Projects#eligibleNewOwners", () => {
+  testGetter(
+    "https://api.osc-fr1.scalingo.com/v1/projects/project-id/eligible_new_owners",
+    null,
+    "eligible_new_owners",
+    (client) => {
+      return new Projects(client).eligibleNewOwners("project-id");
+    },
+  );
+});
+
+describe("Projects#events", () => {
+  testGetter(
+    "https://api.osc-fr1.scalingo.com/v1/projects/project-id/events",
+    null,
+    null,
+    (client) => {
+      return new Projects(client).events("project-id", { page: 1, from: 2 });
+    },
+  );
+
+  testParamsGetter(
+    "https://api.osc-fr1.scalingo.com/v1/projects/project-id/events",
+    { page: 1, per_page: 2, from: 1 },
+    null,
+    (client) => {
+      return new Projects(client).events("project-id", {
+        page: 1,
+        per_page: 2,
+        from: 1,
       });
     },
   );