+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.17.0
+
+* feat: add `ProjectTransferInvitation` interface
+* feat: add project transfer invitations endpoints (index, show, create, accept, decline, cancel)
+* feat: add `in_free_trial` to `AppOwner` interface
+* feat: add events endpoint on `Project`
+* chore: bump dependencies
+
 ## 0.16.1
 
 * feat: add `databases` to `Databases` index